requirements.txt
matplotlib==2.0.2
numpy==1.13.3
tensorflow_gpu==1.4.0
